requirements.txt
# Core Data Processing - Compatible versions with Polars integration
pandas>=2.0.0,<2.1.0
numpy>=1.24.0,<1.25.0
pyspark>=3.5.0,<3.6.0
pyarrow>=7.0.0,<8.0.0  # Fixed: Compatible with Polars requirements

# Polars High-Performance DataFrame Library
polars[pandas,numpy,pyarrow]>=0.20.0,<1.0.0  # Feature flags for interoperability

# Machine Learning
scikit-learn>=1.3.0,<1.4.0
joblib>=1.3.0,<1.4.0

# Advanced ML Models (missing dependencies identified)
# Note: XGBoost >=1.7.0 and LightGBM >=3.3.0 have native Polars DataFrame support
xgboost>=1.7.0,<2.0.0
lightgbm>=3.3.0,<4.0.0  
catboost>=1.2.0,<2.0.0
optuna>=3.0.0,<4.0.0
shap>=0.42.0,<1.0.0

# Data Quality & Validation - Simplified versions
great-expectations>=0.15.0,<0.18.0
pandera>=0.15.0,<0.18.0

# Airflow - Compatible versions to avoid provider conflicts
apache-airflow>=2.6.0,<2.8.0
apache-airflow-providers-google>=8.0.0,<11.0.0

# Jupyter & Analysis
jupyter>=1.0.0
jupyterlab>=3.0.0,<4.0.0
ipykernel>=6.0.0,<7.0.0
matplotlib>=3.7.0,<3.8.0
seaborn>=0.12.0,<0.13.0
plotly>=5.15.0,<5.16.0

# GCP Integration - Compatible versions
google-cloud-core>=2.3.0,<3.0.0
google-cloud-storage>=2.8.0,<2.11.0
google-cloud-bigquery>=3.4.0,<4.0.0  # Updated: Compatible with shapely 2.x

# Utilities
python-dotenv>=1.0.0
click>=8.1.0,<9.0.0
PyYAML>=6.0
pyyaml>=6.0  # Ensure YAML support for configuration

# Data Generation
faker>=18.0.0,<19.0.0

# Code Quality & Testing
black>=23.7.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.4.0,<1.6.0
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0

# Additional utilities for data processing
python-dateutil>=2.8.0,<3.0.0
requests>=2.25.0,<3.0.0